refactor: Utils 메서드 역할 분리 IOUtils에서 UserUtils, HttpUtils 로 역할 분리
